backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain packages - compatible versions
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40
langchain-text-splitters==0.2.4
langchain-openai==0.1.25

openai==1.93.0
tiktoken==0.9.0

sentence-transformers==5.0.0
torch==2.7.1
faiss-cpu==1.7.4

pymupdf==1.24.2
chromadb==0.4.18
numpy==1.26.4
pandas==2.1.3
pyarrow==18.1.0
scikit-learn==1.7.1
scipy==1.11.3
nltk==3.9.1

httpx==0.25.0
websockets==12.0
pydantic==2.11.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
kaggle==1.7.4.5